@@ -1,5 +1,5 @@
 {
-  "name": "t1",
+  "name": "new",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -7,20 +7,21 @@
     "build": "vue-cli-service build"
   },
   "dependencies": {
+    "@tailwindcss/typography": "^0.3.1",
+    "autoprefixer": "^9.8.6",
     "core-js": "^3.6.5",
-    "tailwindcss": "^1.9.6",
-    "vue": "^2.6.11",
-    "vue-router": "^3.2.0",
-    "vue-tailwind": "^1.3.3",
-    "vuex": "^3.4.0"
+    "postcss": "^7.0.35",
+    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.0.1",
+    "vue": "^3.0.0",
+    "vue-router": "^4.0.0-0",
+    "vuex": "^4.0.0-0"
   },
   "devDependencies": {
-    "@tailwindcss/custom-forms": "^0.2.1",
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-router": "~4.5.0",
     "@vue/cli-plugin-vuex": "~4.5.0",
     "@vue/cli-service": "~4.5.0",
-    "vue-template-compiler": "^2.6.11"
+    "@vue/compiler-sfc": "^3.0.0"
   },
   "browserslist": [
     "> 1%",
