@@ -1,22 +1,14 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'
 import store from './store'
-import './css/style.css'
-import VueTailwind from 'vue-tailwind'
-import settings from './settings'
-import components from './components/index'
+import './css/index.css'
+import components from './components'
 
-Vue.config.productionTip = false
-
-Vue.use(VueTailwind, settings)
+const app = createApp(App)
 
 for (let key in components) {
-  Vue.component(key, components[key]);
+    app.component(key, components[key])
 }
 
-new Vue({
-  router,
-  store,
-  render: h => h(App)
-}).$mount('#app')
+app.use(store).use(router).mount('#app')
