@@ -1,9 +1,5 @@
-import Vue from 'vue'
-import VueRouter from 'vue-router'
-import Index from '../layout/Index'
-
-Vue.use(VueRouter)
-
+import { createRouter, createWebHistory } from 'vue-router'
+import Index from '../layout/Index.vue'
 const routes = [
   {
     path: '/',
@@ -12,9 +8,8 @@ const routes = [
   }
 ]
 
-const router = new VueRouter({
-  mode: 'history',
-  base: process.env.BASE_URL,
+const router = createRouter({
+  history: createWebHistory(process.env.BASE_URL),
   routes
 })
 
